+++ sdk/storage/azure-storage-blobs/inc/azure/storage/blobs/blob_options.hpp
@@ -453,95 +453,6 @@ namespace Azure { namespace Storage { namespace Blobs {
   };
 
   /**
-   * @brief Optional parameters for BlobContainerClient::AcquireLease.
-   */
-  struct AcquireBlobContainerLeaseOptions
-  {
-    /**
-     * @brief Context for cancelling long running operations.
-     */
-    Azure::Core::Context Context;
-
-    /**
-     * @brief Optional conditions that must be met to perform this operation.
-     */
-    ModifiedTimeConditions AccessConditions;
-  };
-
-  /**
-   * @brief Optional parameters for BlobContainerClient::RenewLease.
-   */
-  struct RenewBlobContainerLeaseOptions
-  {
-    /**
-     * @brief Context for cancelling long running operations.
-     */
-    Azure::Core::Context Context;
-
-    /**
-     * @brief Optional conditions that must be met to perform this operation.
-     */
-    ModifiedTimeConditions AccessConditions;
-  };
-
-  /**
-   * @brief Optional parameters for BlobContainerClient::ChangeLease.
-   */
-  struct ChangeBlobContainerLeaseOptions
-  {
-    /**
-     * @brief Context for cancelling long running operations.
-     */
-    Azure::Core::Context Context;
-
-    /**
-     * @brief Optional conditions that must be met to perform this operation.
-     */
-    ModifiedTimeConditions AccessConditions;
-  };
-
-  /**
-   * @brief Optional parameters for BlobContainerClient::ReleaseLease.
-   */
-  struct ReleaseBlobContainerLeaseOptions
-  {
-    /**
-     * @brief Context for cancelling long running operations.
-     */
-    Azure::Core::Context Context;
-
-    /**
-     * @brief Optional conditions that must be met to perform this operation.
-     */
-    ModifiedTimeConditions AccessConditions;
-  };
-
-  /**
-   * @brief Optional parameters for BlobContainerClient::BreakLease.
-   */
-  struct BreakBlobContainerLeaseOptions
-  {
-    /**
-     * @brief Context for cancelling long running operations.
-     */
-    Azure::Core::Context Context;
-
-    /**
-     * @brief Proposed duration the lease should continue before it is broken, in seconds,
-     * between 0 and 60. This break period is only used if it is shorter than the time remaining on
-     * the lease. If longer, the time remaining on the lease is used. A new lease will not be
-     * available before the break period has expired, but the lease may be held for longer than the
-     * break period.
-     */
-    Azure::Core::Nullable<int32_t> BreakPeriod;
-
-    /**
-     * @brief Optional conditions that must be met to perform this operation.
-     */
-    ModifiedTimeConditions AccessConditions;
-  };
-
-  /**
    * @brief Optional parameters for BlobClient::GetProperties.
    */
   struct GetBlobPropertiesOptions
@@ -783,7 +694,7 @@ namespace Azure { namespace Storage { namespace Blobs {
   };
 
   /**
-   * @brief Optional parameters for BlobClient::AcquireLease.
+   * @brief Optional parameters for BlobLeaseClient::Acquire.
    */
   struct AcquireBlobLeaseOptions
   {
@@ -799,7 +710,7 @@ namespace Azure { namespace Storage { namespace Blobs {
   };
 
   /**
-   * @brief Optional parameters for BlobClient::RenewLease.
+   * @brief Optional parameters for BlobLeaseClient::Renew.
    */
   struct RenewBlobLeaseOptions
   {
@@ -815,7 +726,7 @@ namespace Azure { namespace Storage { namespace Blobs {
   };
 
   /**
-   * @brief Optional parameters for BlobClient::ChangeLease.
+   * @brief Optional parameters for BlobLeaseClient::Change.
    */
   struct ChangeBlobLeaseOptions
   {
@@ -831,7 +742,7 @@ namespace Azure { namespace Storage { namespace Blobs {
   };
 
   /**
-   * @brief Optional parameters for BlobClient::ReleaseLease.
+   * @brief Optional parameters for BlobLeaseClient::Release.
    */
   struct ReleaseBlobLeaseOptions
   {
@@ -847,7 +758,7 @@ namespace Azure { namespace Storage { namespace Blobs {
   };
 
   /**
-   * @brief Optional parameters for BlobClient::BreakLease.
+   * @brief Optional parameters for BlobLeaseClient::Break.
    */
   struct BreakBlobLeaseOptions
   {